requirements.txt
urwid
pysrt
numpy
pillow==2.8.2
twython
python-tumblpy
